refactor: move CSV helper functions to a separate helpers.js file
